Include in CI - Dockerize the client
The CI environment depends on docker for building the client
and for the running of style and lint checks.

The first step to include the client in the CI environment
is thus to dockerize the client.

Docker won't allow including files from higher directories
The requirements.txt file is therefore moved to the src directory