Include in CI - Jenkinsfile
The pipeline defined in the Jenkinsfile relies on the
dockerized client.

The client is build and then the tests (which are not yet present),
will run. The build and tests are initiated from the docker-compose
in the src/jenkins folder

If the build and the tests all finish succesfully the client
is ready for integration