Add todo for type conversion
Type conversion has been changed to be data driven. This requires
however additional knowledge which would otherwise have been
programmed. An example is the bool conversion that is now
programmed as N = False, else is True. This is however very
specific for the import and should not be coded like this.