Remove shared code from project
Typeconversion and logging have been moved the GOB-Core project and are
imported here. Connection to the messagebroker had been moved to the
GOB-Core before, it has been removed here.

This commit leaves the project in an indetermined state, because also
the separation of mapping and model needs to bee done also.