Use volgnummer from column on GGW GGP
Changing volgnummer on GGW/GGP resulted in some findings on entity validation and convertion. The code assumed an entity with states always has a column named ‘volgnummer’ while this should be configurable in the import mapping. Fox example the column in the GGW/GGP files was named ‘VOLGNUMMER’. Changes were made to core and validator to make this possible.